+++ CHANGELOG.md
@@ -8,6 +8,16 @@ Operators, you should copy/paste content of this content straight to your projec
 
 If you were at `firehose-core` version `1.0.0` and are bumping to `1.1.0`, you should copy the content between those 2 version to your own repository, replacing placeholder value `fire{chain}` with your chain's own binary.
 
+## Unreleased
+
+### Substreams
+
+* Add shared cache for tier1 execution near HEAD, to prevent multiple tier1 instances from reprocessing the same module on the same block when it comes in (ex: foundational modules)
+
+### Tools
+
+* make 'compare-blocks' command support one-blocks stores as well as merged-blocks
+
 ## v1.7.4
 
 - Bump `substreams` lib to `v1.12.3`